small fixes

+++ leetcode/easy/missing_number/main.cpp
@@ -6,13 +6,13 @@
 
 using namespace std;
 
-int missing_number(vector<int>& nums) {
+int missing_number(vector<int> &nums) {
   sort(nums.begin(), nums.end());
 
   if (nums.front() != 0) {
     return 0;
   }
-  if (nums.back() != nums.size()) {
+  if (nums.back() != static_cast<int>(nums.size())) {
     return nums.size();
   }
 
@@ -27,7 +27,7 @@ int missing_number(vector<int>& nums) {
 }
 
 // another approach
-int missing_number2(vector<int>& nums) {
+int missing_number2(vector<int> &nums) {
   set<int> nums_set(nums.begin(), nums.end());
 
   for (auto i = 0u; i < nums.size() + 1; ++i) {
@@ -39,7 +39,7 @@ int missing_number2(vector<int>& nums) {
 }
 
 // approach with gauss sum
-int missing_number3(vector<int>& nums) {
+int missing_number3(vector<int> &nums) {
   const auto expected_sum = nums.size() * (nums.size() + 1) / 2;
   const auto real_sum = accumulate(nums.begin(), nums.end(), 0);
   return expected_sum - real_sum;
